backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
captum==0.6.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
reportlab==4.0.7
sqlalchemy==2.0.23
pydantic==2.5.0
aiofiles==23.2.1
# Additional dependencies for training
datasets==2.14.0
accelerate==0.24.0
tensorboard==2.15.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
